moved all HDC code to upload_hdc
