+++ templates/actor-sidebar/actor-sidebar-sheet.hbs
@@ -70,7 +70,7 @@
 
             <label class="upload-button">
                 <span>Upload .HDC</span>
-                <input type="file" />
+                <input type="file" accept=".hdc" />
             </label>
         </div>
     </header>
